pyproject.toml
[project]
name = "fast-api1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.0.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "fastapi-filter[sqlalchemy]>=2.0.1",
    "fastapi-pagination>=0.13.3",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "openpyxl>=3.1.5",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.0.0",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.4",
    "sqlalchemy[asyncio]>=2.0.43",
    "taskiq>=0.11.20",
    "taskiq-aio-pika>=0.4.4",
    "taskiq-fastapi>=0.3.5",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
